fix: remove null accordions from accordionlist
It may happen that an accordionlist is generated in react by using
the notation `data && <Accordion...> and where data is null it would
result in a null accordion.

The expected result would be to not have that specific Accordion, but
instead it fails.

This commit addresses this problem by removing all falsey values from
the accordion list